Remove NavBar to src and rename it to Layout
