+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import Layout from './Layout';
 import Calculator from './components/Calculator';
-import NavBar from './components/NavBar';
 import Home from './components/Home';
 import Quote from './components/Quote';
 
@@ -11,7 +11,7 @@ function App() {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<NavBar />}>
+        <Route path="/" element={<Layout />}>
           <Route index element={<Home />} />
           <Route path="calculator" element={<Calculator />} />
           <Route path="quote" element={<Quote />} />
